+++ tests/PSRule.Tests/AssertTests.cs
@@ -32,6 +32,20 @@ namespace PSRule
         }
 
         [Fact]
+        public void WithinRollupBlock()
+        {
+            SetContext();
+            var assert = GetAssertionHelper();
+            var actual1 = PSRule.Runtime.RuleConditionResult.Create(new object[] { PSObject.AsPSObject(assert.Create(true, "Test reason")), PSObject.AsPSObject(assert.Create(false, "Test reason")) });
+            Assert.True(actual1.AnyOf());
+            Assert.False(actual1.AllOf());
+
+            var actual2 = PSRule.Runtime.RuleConditionResult.Create(new object[] { assert.Create(true, "Test reason"), assert.Create(false, "Test reason") });
+            Assert.True(actual2.AnyOf());
+            Assert.False(actual2.AllOf());
+        }
+
+        [Fact]
         public void HasJsonSchema()
         {
             SetContext();
@@ -335,6 +349,7 @@ namespace PSRule
         {
             var context = PipelineContext.New(new Configuration.PSRuleOption(), null, null, null, null);
             context.ExecutionScope = ExecutionScope.Condition;
+            new RunspaceContext(context, null);
         }
 
         private static PSObject GetObject(params (string name, object value)[] properties)